test: split client_test.go into focused files
- client_helpers_test.go: Test helper functions and mock daemon setup
- client_connection_test.go: Client creation and connection tests
- client_subscribe_test.go: Subscribe functionality tests
- client_send_test.go: SendEvent functionality tests
- client_close_test.go: Close, notify callback, and error handling tests
- client_deadline_test.go: Write deadline regression tests
- client_backpressure_test.go: Queue backpressure tests
- client_reconnection_test.go: Reconnection and exponential backoff tests
- client_batching_test.go: Event batching during network failure tests
- client_receiver_test.go: Event receiver loop error handling tests
- client_disconnect_test.go: Disconnect detection tests
